refactor : 레이아웃 폴더 구조 변경
