+++ client/user/src/app/_components/Layout/CommonLayout.tsx
@@ -1,0 +1,13 @@
+import { Footer, Header } from '@/components'
+
+export const CommonLayout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <div className="flex min-h-screen flex-col">
+      <Header />
+      <main className="flex-1 overflow-hidden">{children}</main>
+      <div className="mt-auto">
+        <Footer />
+      </div>
+    </div>
+  )
+}
