+++ client/user/src/app/_components/Layout/Meta.tsx
@@ -1,0 +1,14 @@
+import { GoogleTagManager } from '@/components'
+import Head from 'next/head'
+
+export const Meta = () => {
+  return (
+    <Head>
+      <GoogleTagManager />
+      <link rel="icon" type="image/png" href="/images/Logo.png" />
+      <meta property="og:image" content="/images/Logo.png" />
+      <meta property="og:url" content="https://tookscan.com" />
+      <meta property="og:locale" content="ko_KR" />
+    </Head>
+  )
+}
