+++ client/user/src/app/_components/index.ts
@@ -1,9 +1,10 @@
+export * from './GradientCard'
+export * from './Layout'
+export * from './Modal'
+export * from './ReviewCard'
 export * from './ScrollPages/1stPage'
 export * from './ScrollPages/2stPage'
 export * from './ScrollPages/3stPage'
 export * from './ScrollPages/4stPage'
 export * from './ScrollPages/5stPage'
-export * from './GradientCard'
-export * from './Modal'
-export * from './ReviewCard'
 export * from './SmallCard'
