+++ client/user/src/app/layout.tsx
@@ -1,47 +1,20 @@
-import {
-  Footer,
-  GoogleTagManager,
-  GoogleTagManagerNoScript,
-  Header,
-} from '@/components'
+'use client'
+
+import { CommonLayout, Meta, Providers } from '@/app/_components'
+import { GoogleTagManagerNoScript } from '@/components'
 import type { LayoutProps } from '@/types/common'
-import {
-  CommonSpriteSheet,
-  Modal,
-  QueryProvider,
-  SpriteSheet,
-  ToastProvider,
-} from '@tookscan/components'
 import '@tookscan/styles/globals.css'
 import '@tookscan/styles/reset.css'
-import Head from 'next/head'
 
 export default function RootLayout({ children }: LayoutProps) {
   return (
     <html lang="ko">
-      <Head>
-        <GoogleTagManager />
-        <link rel="icon" type="image/png" href="/images/Logo.png" />
-        <meta property="og:image" content="/images/Logo.png" />
-        <meta property="og:url" content="https://tookscan.com" />
-        <meta property="og:locale" content="ko_KR" />
-      </Head>
+      <Meta />
       <body>
-        {SpriteSheet}
-        {CommonSpriteSheet}
         <GoogleTagManagerNoScript />
-        <QueryProvider>
-          <Modal />
-          <ToastProvider>
-            <div className="flex min-h-screen flex-col">
-              <Header />
-              <main className="flex-1 overflow-hidden">{children}</main>
-              <div className="mt-auto">
-                <Footer />
-              </div>
-            </div>
-          </ToastProvider>
-        </QueryProvider>
+        <Providers>
+          <CommonLayout>{children}</CommonLayout>
+        </Providers>
       </body>
     </html>
   )
