+++ packages/components/ui/Modal/index.ts
@@ -1,4 +1,4 @@
 export * from './Confirm'
-export * from './Modal'
+export * from './ModalSheet'
 export * from './SearchAddress'
-export * from './Toast'
+export * from './ToastSheet'
