+++ packages/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useAuth'
 export * from './useModal'
 export * from './usePhoneAuth'
+export * from './useToast'
